+++ unix/Xvnc/programs/Xserver/hw/vnc/auth.c
@@ -5,7 +5,7 @@
  */
 
 /*
- *  Copyright (C) 2010, 2012-2020 D. R. Commander.  All Rights Reserved.
+ *  Copyright (C) 2010, 2012-2021 D. R. Commander.  All Rights Reserved.
  *  Copyright (C) 2010 University Corporation for Atmospheric Research.
  *                     All Rights Reserved.
  *  Copyright (C) 2003-2006 Constantin Kaplinsky.  All Rights Reserved.
@@ -1511,7 +1511,7 @@ void rfbVncAuthProcessResponse(rfbClientPtr cl)
     rfbClientPtr otherCl;
 
     rfbAuthUnblock(cl->host);
-    if (!cl->reverseConnection && rfbNeverShared && rfbDontDisconnect) {
+    if (!cl->reverseConnection && rfbNeverShared && !rfbDisconnect) {
       for (otherCl = rfbClientHead; otherCl; otherCl = otherCl->next) {
         if ((otherCl != cl) && (otherCl->state == RFB_NORMAL)) {
           rfbLog("-dontdisconnect: Not shared & existing client\n");